+++ src/main/java/com/example/HumanResourceApplication/repository/EmployeeRepository.java
@@ -17,20 +17,20 @@ import java.util.Optional;
 
 
 @RepositoryRestResource(path = "employees")
-public interface EmployeeRepository extends JpaRepository<Employee, Double> {
+public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
 
-    Optional<Employee> findByEmployeeId(Double empId);
+    Optional<Employee> findByEmployeeId(Integer empId);
 
     List<ManagerProjection> findDistinctBySubordinatesIsNotEmpty();
 
     //List<EmployeeProjection>findByManagerId(Long Id);
 
-    List<EmployeeProjection> findByManager_EmployeeId(Double id);
+    List<EmployeeProjection> findByManager_EmployeeId(Integer id);
     @RestResource(path = "by-email", rel = "managerByEmail")
     ManagerProjection findDistinctBySubordinatesIsNotEmptyAndEmail(String email);
 
 
-    default List<Employee> getHierarchy(Double id) {
+    default List<Employee> getHierarchy(Integer id) {
         List<Employee> hierarchy = new ArrayList<>();
 
         Employee current = findByEmployeeId(id)
@@ -48,6 +48,6 @@ public interface EmployeeRepository extends JpaRepository<Employee, Double> {
     List<ManagerProjection> findDistinctBySubordinatesIsNotEmptyAndDepartment_DepartmentName(String departmentName);
 
 
-    long countByManager_EmployeeId(Double managerId);
+    long countByManager_EmployeeId(Integer managerId);
 
 }
